requirements.txt
fastapi
uvicorn
openai
pinecone
pydantic
python-dotenv

# For document parsing and file uploads
pypdf
python-docx
python-multipart